requirements.txt
aiogram==3.4.1
aiohttp==3.9.5
aiosqlite==0.19.0
python-dotenv==1.0.1
certifi==2024.2.2
